refactor: LottoMatcher의 역할을 LottoGame과 LottoResult로 분리
